Exterminate all warnings

+++ Src/Graphics/GraphicsDX11.cpp
@@ -268,14 +268,14 @@ void GraphicsDX11::swap() {
 
 void GraphicsDX11::clear(Color color) {
     float clearColor[4] = {color.red,color.green,color.blue,color.alpha};
-    for (int i=0;i<currentRenderTargetViews.size();i++) {
+    for (int i = 0; i < (int)currentRenderTargetViews.size(); i++) {
         dxContext->ClearRenderTargetView( currentRenderTargetViews[i], clearColor );
     }
     dxContext->ClearDepthStencilView( currentDepthStencilView, D3D11_CLEAR_DEPTH|D3D11_CLEAR_STENCIL, 1.f, 0 );
 }
 
 void GraphicsDX11::setRenderTarget(Texture* renderTarget) {
-    for (int i=0;i<currentRenderTargetViews.size();i++) {
+    for (int i = 0; i < (int)currentRenderTargetViews.size(); i++) {
         currentRenderTargetViews[i] = nullptr;
     }
     dxContext->OMSetRenderTargets( currentRenderTargetViews.size(), currentRenderTargetViews.data(), nullptr );
@@ -286,14 +286,14 @@ void GraphicsDX11::setRenderTarget(Texture* renderTarget) {
 }
 
 void GraphicsDX11::setRenderTargets(std::vector<Texture*> renderTargets) {
-    for (int i=0;i<currentRenderTargetViews.size();i++) {
+    for (int i = 0; i < (int)currentRenderTargetViews.size(); i++) {
         currentRenderTargetViews[i] = nullptr;
     }
     dxContext->OMSetRenderTargets( currentRenderTargetViews.size(), currentRenderTargetViews.data(), nullptr );
 
     currentRenderTargetViews.clear();
     TextureDX11* maxSizeTexture = (TextureDX11*)renderTargets[0];
-    for (int i=0;i<renderTargets.size();i++) {
+    for (int i = 0; i < (int)renderTargets.size(); i++) {
         if (!renderTargets[i]->isRenderTarget()) {
             throwException("setRenderTargets","renderTargets["+String::fromInt(i)+"] is not a valid render target");
         }
@@ -302,7 +302,7 @@ void GraphicsDX11::setRenderTargets(std::vector<Texture*> renderTargets) {
             maxSizeTexture = (TextureDX11*)renderTargets[i];
         }
     }
-    for (int i=0;i<renderTargets.size();i++) {
+    for (int i = 0; i < (int)renderTargets.size(); i++) {
         if (renderTargets[i]->getWidth()>maxSizeTexture->getWidth() || renderTargets[i]->getHeight()>maxSizeTexture->getHeight()) {
             throwException("setRenderTargets",
                 "Render target sizes are incompatible ("+String::fromInt(maxSizeTexture->getWidth())+","+String::fromInt(maxSizeTexture->getHeight())+" vs "+
@@ -314,7 +314,7 @@ void GraphicsDX11::setRenderTargets(std::vector<Texture*> renderTargets) {
 }
 
 void GraphicsDX11::resetRenderTarget() {
-    for (int i=0;i<currentRenderTargetViews.size();i++) {
+    for (int i = 0; i < (int)currentRenderTargetViews.size(); i++) {
         currentRenderTargetViews[i] = nullptr;
     }
     dxContext->OMSetRenderTargets( currentRenderTargetViews.size(), currentRenderTargetViews.data(), nullptr );
@@ -332,8 +332,8 @@ void GraphicsDX11::setViewport(Rectanglei vp) {
         dxViewport.Height = (FLOAT)viewport.height();
         dxViewport.MinDepth = 0.0f;
         dxViewport.MaxDepth = 1.0f;
-        dxViewport.TopLeftX = viewport.topLeftCorner().x;
-        dxViewport.TopLeftY = viewport.topLeftCorner().y;
+        dxViewport.TopLeftX = (FLOAT)viewport.topLeftCorner().x;
+        dxViewport.TopLeftY = (FLOAT)viewport.topLeftCorner().y;
         dxContext->RSSetViewports( 1, &dxViewport );
     }
 }
